:sparkles: :construction: add tags in product

+++ app/src/main/java/com/ezlinker/app/modules/tag/model/Tag.java
@@ -30,7 +30,7 @@ public class Tag extends XEntity {
     /**
      * Tag 作用的对象ID
      */
-    private Integer linkId;
+    private Long linkId;
 
 
 }
